refactor: 친구 구조 리팩토링
