+++ src/main/java/io/github/eappezo/soundary/core/persistence/infrastructure/FriendEntity.java
@@ -1,7 +1,5 @@
-package io.github.eappezo.soundary.services.friend.domain;
+package io.github.eappezo.soundary.core.persistence.infrastructure;
 
-import io.github.eappezo.soundary.core.persistence.infrastructure.BaseEntity;
-import io.github.eappezo.soundary.services.friend.domain.key.FriendKey;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.IdClass;
@@ -9,12 +7,12 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Entity
+@Entity(name = "friends")
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @IdClass(FriendKey.class)
-public class Friend extends BaseEntity {
+public class FriendEntity extends BaseEntity {
     @Id
     private String fromUserId;
