+++ src/main/java/io/github/eappezo/soundary/core/persistence/infrastructure/FriendKey.java
@@ -1,6 +1,8 @@
-package io.github.eappezo.soundary.services.friend.domain.key;
+package io.github.eappezo.soundary.core.persistence.infrastructure;
 
 import java.io.Serializable;
+
+import io.github.eappezo.soundary.core.identification.Identifier;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -13,4 +15,8 @@ import lombok.NoArgsConstructor;
 public class FriendKey implements Serializable {
     private String fromUserId;
     private String toUserId;
+
+    public static FriendKey of(Identifier fromUserId, Identifier toUserId) {
+        return new FriendKey(fromUserId.toString(), toUserId.toString());
+    }
 }
