+++ src/main/java/io/github/eappezo/soundary/services/friend/application/FriendRepository.java
@@ -1,13 +1,13 @@
 package io.github.eappezo.soundary.services.friend.application;
 
-import io.github.eappezo.soundary.services.friend.domain.Friend;
-import io.github.eappezo.soundary.services.friend.domain.key.FriendKey;
+import io.github.eappezo.soundary.core.persistence.infrastructure.FriendEntity;
+import io.github.eappezo.soundary.core.persistence.infrastructure.FriendKey;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FriendRepository extends JpaRepository<Friend, FriendKey> {
+public interface FriendRepository extends JpaRepository<FriendEntity, FriendKey> {
 
-    List<Friend> findByFromUserId(String fromUserId);
+    List<FriendEntity> findByFromUserId(String fromUserId);
 
-    List<Friend> findByToUserId(String toUserId);
+    List<FriendEntity> findByToUserId(String toUserId);
 }
